+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 ### Bug fixes
 - Allow `np.bool_` as a valid `bool` dtype when validating. @dsleiter (#505)
 
+### New features
+- `GroupValidator` now checks if child groups, datasets, and links have the correct quantity of elements and returns
+  an `IncorrectQuantityError` for each mismatch. @dsleiter (#500)
+
 ## HDMF 2.3.0 (December 8, 2020)
 
 ### New features