+++ CHANGELOG.md
@@ -9,6 +9,10 @@
 - `GroupValidator` now checks if child groups, datasets, and links have the correct quantity of elements and returns
   an `IncorrectQuantityError` for each mismatch. @dsleiter (#500)
 
+### Bug fixes
+- Fix building of Data objects where the spec has no dtype and the Data object value is a DataIO wrapping an
+  AbstractDataChunkIterator. @rly (#512)
+
 ## HDMF 2.3.0 (December 8, 2020)
 
 ### New features