+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # HDMF Changelog
 
+## HDMF 2.4.0 (Upcoming)
+
+### Internal improvements
+- Update CI. @rly (#432)
+
+### Bug fixes
+- Allow `np.bool_` as a valid `bool` dtype when validating. @dsleiter (#505)
+
+### New features
+- `GroupValidator` now checks if child groups, datasets, and links have the correct quantity of elements and returns
+  an `IncorrectQuantityError` for each mismatch. @dsleiter (#500)
+
+### Bug fixes
+- Fix building of Data objects where the spec has no dtype and the Data object value is a DataIO wrapping an
+  AbstractDataChunkIterator. @rly (#512)
+
 ## HDMF 2.3.0 (December 8, 2020)
 
 ### New features